Reorganize components directory with domain-driven architecture
Major restructuring to improve code organization, maintainability, and developer experience:

STRUCTURE CHANGES:
- Create domain-based folder hierarchy: common/ui, features/, layout/, visualization/
- Group components by business purpose rather than technical type
- Separate reusable UI primitives from feature-specific components

COMPONENT MOVES:
- UI components → common/ui/ (avatar, textarea)
- Feature components → features/{auth,chat,dashboard,landing}/
- Layout components → layout/ (dashboard-sidebar)
- 3D/shader components → visualization/{3d,shaders}/
- Keep icons/ at root for atomic components

IMPORT UPDATES:
- Update 27+ import statements across app, components, and tests
- Fix dynamic imports in dashboard preload hook
- Maintain all existing functionality and behavior

CLEANUP:
- Remove empty loader/ directory
- Delete unused v0-chat-demo.tsx wrapper
- Preserve all component functionality

Benefits: Better scalability, clearer code ownership, faster development, easier onboarding
