+++ src/app/dashboard/chat/page.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { DashboardPageWrapper } from "@/components/dashboard-page-wrapper"
+import { DashboardPageWrapper } from "@/components/features/dashboard/dashboard-page-wrapper"
 import { CONTAINER_WIDTHS } from "@/constants/layout"
 
 export default function Chat() {
