+++ src/app/page.tsx
@@ -1,8 +1,8 @@
 "use client"
 
-import Header from "@/components/header"
-import HeroContent from "@/components/hero-content"
-import ShaderBackground from "@/components/shader-background"
+import Header from "@/components/features/landing/header"
+import HeroContent from "@/components/features/landing/hero-content"
+import ShaderBackground from "@/components/visualization/shaders/shader-background"
 
 export default function ShaderShowcase() {
   return (
