+++ src/components/features/chat/v0-ai-chat.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useRef, useCallback } from "react";
 import { useState } from "react";
-import { Textarea } from "@/components/ui/textarea";
+import { Textarea } from "@/components/common/ui/textarea";
 import { cn } from "@/lib/utils";
 import {
     ImageIcon,
