+++ src/components/features/landing/hero-content.tsx
