+++ src/components/layout/dashboard-sidebar.tsx
@@ -5,7 +5,7 @@ import { usePathname } from "next/navigation"
 import { useUser } from "@stackframe/stack"
 import { useState } from "react"
 import { UserRoundIcon } from "lucide-react"
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
+import { Avatar, AvatarFallback, AvatarImage } from "@/components/common/ui/avatar"
 import { ClockIcon } from "@/components/icons/clock-icon"
 import { MessageSquareIcon } from "@/components/icons/message-square-icon"
 import { SparklesIcon } from "@/components/icons/sparkles-icon"
