--- src/components/v0-chat-demo.tsx
@@ -1,5 +1,0 @@
-import { VercelV0Chat } from "@/components/ui/v0-ai-chat"
-
-export function Demo() {
-    return <VercelV0Chat />
-}
