+++ src/components/visualization/3d/sphere-3d-loading-skeleton.tsx
