+++ src/components/visualization/shaders/shader-skeleton.tsx
